+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- [BREAKING] Changed minimum Elixir version from 1.16 to 1.18
+
+### Removed
+- Removed `Jason` dependency
 
 ## [0.4.1] - 2025-09-23
 ### Fixed