+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - Removed `Jason` dependency
+- Removed deprecated `Tornex.API.torn_get/1`
 
 ## [0.4.1] - 2025-09-23
 ### Fixed