+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.0] - 2026-03-31
 ### Added
 - Added per-node ratelimiting
 - Added combinatorial queries to combine similar `Tornex.Scheduler.SpecQuery` into one `Tornex.Scheduler.ExecutionUnit` to be performed as one API call.